MXS-5187 Set current_gtid_list at end of transaction
Update m_current_gtid_list when the transaction commits,
remove mutex as the variable is only accessed internally
(which includes FileWriter).

This has no effect on functionality, but in the next steps
there is a gtid being received and the current gtid, which
has the same meaning as elsewhere, i.e. a committed trx.